Sync failures
The names used in plugins.js got out of sync with the names in
failures.js.

Also, removed "Plugin" and "PluginFactory" and "Has" from
failure names, since they were useless words.

--

Renamed `Plugin` to `pluginFactory` in plugins.js.

This is entirely an internal renaming. Meant to have it in its own commit. :(